Burden: Remove `literature` from required field
The publication supporting Genebass associations hasn't been peer reviewed yet, therefore we don't have a suitable ID for it.
As discussed, we want to include an additional field in the schema to accommodate PPR IDs - especially to get EPMC evidence, however we don't want to do this right now as it would imply changes in the API and so on. So for the time being, Genebass evidence will only be referenced with a linkout to their app and not with a publication.